Fix: 웹 소켓 연결 시 메세지를 제대로 받고 보내도록 수정
