+++ src/main/java/com/knu/KnowcKKnowcK/domain/MessageThread.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 
 import java.time.LocalDateTime;
 
@@ -18,11 +20,13 @@ public class MessageThread {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long messageThreadId;
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="message_id")
+    @OnDelete(action = OnDeleteAction.CASCADE)
     private Message messageId;
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="member_id")
+    @OnDelete(action = OnDeleteAction.CASCADE)
     private Member memberId;
 
     private String content;
