+++ src/main/java/com/knu/KnowcKKnowcK/dto/requestdto/MessageRequestDTO.java
@@ -2,6 +2,7 @@ package com.knu.KnowcKKnowcK.dto.requestdto;
 
 
 import com.knu.KnowcKKnowcK.domain.DebateRoom;
+import com.knu.KnowcKKnowcK.domain.Member;
 import com.knu.KnowcKKnowcK.domain.Message;
 import lombok.Data;
 
@@ -12,8 +13,9 @@ public class MessageRequestDTO {
     Long roomId;
     String content;
 
-    public Message toMessage(DebateRoom debateRoom){
+    public Message toMessage( Member member, DebateRoom debateRoom){
         return Message.builder()
+                .memberId(member)
                 .debateRoomId(debateRoom)
                 .content(content)
                 .createdTime(LocalDateTime.now())
